+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 0.0.2
 
-* TODO: Describe initial release.
+* first build with bottom sheet.
+
+## 0.0.3
+
+* enable/disable fixed.